+++ homebytwo/routes/tests/test_tasks.py
@@ -1,18 +1,10 @@
 from datetime import timedelta
 
 from homebytwo.routes.models import WebhookTransaction
-from homebytwo.routes.tasks import (
-    import_strava_activities_streams_task,
-    import_strava_activities_task,
-    import_strava_activity_streams_task,
-    process_strava_events,
-    train_prediction_models_task,
-)
-from homebytwo.routes.tests.factories import (
-    ActivityFactory,
-    ActivityTypeFactory,
-    WebhookTransactionFactory,
-)
+from homebytwo.routes.tasks import (import_strava_activities_streams_task, import_strava_activities_task,
+                                    import_strava_activity_streams_task, process_strava_events,
+                                    train_prediction_models_task)
+from homebytwo.routes.tests.factories import ActivityFactory, ActivityTypeFactory, WebhookTransactionFactory
 
 STRAVA_API_BASE_URL = "https://www.strava.com/api/v3/"
 STRAVA_STREAMS_URL = (
@@ -21,12 +13,12 @@ STRAVA_STREAMS_URL = (
 GARMIN_UPLOAD_URL = "https://connect.garmin.com/modern/proxy/upload-service/upload/gpx"
 
 
-def test_import_strava_activities_task(athlete, intercept):
+def test_import_strava_activities_task(athlete, mock_call_json_response):
     url = STRAVA_API_BASE_URL + "athlete/activities"
     call = import_strava_activities_task
     response_json = "activities.json"
 
-    intercept(call, url, response_json, athlete_id=athlete.id)
+    mock_call_json_response(call, url, response_json, athlete_id=athlete.id)
 
     assert athlete.activities.count() == 2
     assert athlete.activities_imported
@@ -51,13 +43,13 @@ def test_import_strava_activities_streams_task(athlete, mocker):
     mock_task.assert_called
 
 
-def test_import_strava_activity_streams_task_success(athlete, intercept):
+def test_import_strava_activity_streams_task_success(athlete, mock_call_json_response):
     activity = ActivityFactory(athlete=athlete, streams=None)
     url = STRAVA_STREAMS_URL.format(activity.strava_id)
     expected = "Streams successfully imported for activity {}".format(
         activity.strava_id
     )
-    response = intercept(
+    response = mock_call_json_response(
         import_strava_activity_streams_task,
         url,
         "streams.json",
@@ -76,12 +68,12 @@ def test_update_activity_streams_from_strava_skip(athlete):
     assert expected in response
 
 
-def test_import_strava_activity_streams_task_missing(athlete, intercept):
+def test_import_strava_activity_streams_task_missing(athlete, mock_call_json_response):
     activity = ActivityFactory(athlete=athlete, streams=None)
     url = STRAVA_STREAMS_URL.format(str(activity.strava_id))
 
     expected = "Streams not imported for activity {}".format(activity.strava_id)
-    response = intercept(
+    response = mock_call_json_response(
         import_strava_activity_streams_task,
         url,
         "missing_streams.json",
@@ -154,7 +146,7 @@ def test_train_prediction_models_task_no_activity(athlete):
     assert expected in response
 
 
-def test_process_strava_events_create_update_delete(athlete, intercept):
+def test_process_strava_events_create_update_delete(athlete, mock_call_json_response):
     activity_strava_id = 1234567890
     WebhookTransactionFactory(
         action="create",
@@ -164,7 +156,7 @@ def test_process_strava_events_create_update_delete(athlete, intercept):
     call = process_strava_events
     url = STRAVA_API_BASE_URL + "activities/" + str(activity_strava_id)
     response_json = "race_run_activity.json"
-    intercept(call, url, response_json)
+    mock_call_json_response(call, url, response_json)
 
     transactions = WebhookTransaction.objects.all()
     processed_transactions = transactions.filter(status=WebhookTransaction.PROCESSED)
@@ -178,7 +170,7 @@ def test_process_strava_events_create_update_delete(athlete, intercept):
         athlete_strava_id=athlete.strava_id,
         activity_strava_id=activity_strava_id,
     )
-    intercept(call, url, response_json)
+    process_strava_events()
     assert processed_transactions.count() == 2
     assert athlete.activities.count() == 1
 
@@ -187,7 +179,7 @@ def test_process_strava_events_create_update_delete(athlete, intercept):
         athlete_strava_id=athlete.strava_id,
         activity_strava_id=activity_strava_id,
     )
-    call()
+    process_strava_events()
     assert processed_transactions.count() == 3
     assert athlete.activities.count() == 0